--- app/src/debug/java/org/sopt/official/FlipperInitializer.kt
@@ -1,50 +1,0 @@
-/*
- * MIT License
- * Copyright 2023 SOPT - Shout Our Passion Together
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-package org.sopt.official
-
-import android.app.Application
-import com.facebook.flipper.android.AndroidFlipperClient
-import com.facebook.flipper.plugins.inspector.DescriptorMapping
-import com.facebook.flipper.plugins.inspector.InspectorFlipperPlugin
-import com.facebook.flipper.plugins.network.FlipperOkhttpInterceptor
-import com.facebook.flipper.plugins.network.NetworkFlipperPlugin
-import com.facebook.soloader.SoLoader
-import okhttp3.OkHttpClient
-
-object FlipperInitializer {
-    private val flipperNetworkPlugin = NetworkFlipperPlugin()
-
-    fun init(app: Application) {
-        SoLoader.init(app, false)
-        val client = AndroidFlipperClient.getInstance(app)
-        client.addPlugin(InspectorFlipperPlugin(app, DescriptorMapping.withDefaults()))
-        client.addPlugin(flipperNetworkPlugin)
-        client.start()
-    }
-
-    fun addFlipperNetworkPlugin(builder: OkHttpClient.Builder): OkHttpClient.Builder {
-        return builder.addNetworkInterceptor(FlipperOkhttpInterceptor(flipperNetworkPlugin))
-    }
-}
