+++ app/src/main/java/org/sopt/official/di/AuthModule.kt
@@ -38,8 +38,6 @@ import org.sopt.official.data.source.api.auth.RemoteAuthDataSource
 import org.sopt.official.data.source.impl.DefaultLocalAuthDataSource
 import org.sopt.official.data.source.impl.DefaultRemoteAuthDataSource
 import org.sopt.official.domain.repository.AuthRepository
-import org.sopt.official.feature.auth.ProvidesIntent
-import org.sopt.official.network.authenticator.ProvidesIntents
 import retrofit2.Retrofit
 import javax.inject.Singleton
 
@@ -71,8 +69,5 @@ object AuthModule {
         @Singleton
         fun bindLocalAuthDataSource(dataSource: DefaultLocalAuthDataSource): LocalAuthDataSource
 
-        @Binds
-        @Singleton
-        fun bindIntent(providesIntents: ProvidesIntent): ProvidesIntents
     }
 }
