+++ app/src/main/java/org/sopt/official/feature/navigator/NavigatorProvidesIntent.kt
@@ -1,14 +1,15 @@
-package org.sopt.official.feature.auth
+package org.sopt.official.feature.navigator
 
 import android.content.Context
 import android.content.Intent
 import dagger.hilt.android.qualifiers.ApplicationContext
-import org.sopt.official.network.authenticator.ProvidesIntents
+import org.sopt.official.common.navigator.NavigatorProvides
+import org.sopt.official.feature.auth.AuthActivity
 import javax.inject.Inject
 
-class ProvidesIntent @Inject constructor(
+class NavigatorProvidesIntent @Inject constructor(
     @ApplicationContext private val context: Context
-) : ProvidesIntents {
+) : NavigatorProvides {
     override fun getAuthActivityIntent(): Intent =
         AuthActivity.newInstance(context)
