+++ core/common/src/main/java/org/sopt/official/common/navigator/NavigatorProvides.kt
@@ -1,0 +1,7 @@
+package org.sopt.official.common.navigator
+
+import android.content.Intent
+
+interface NavigatorProvides {
+    fun getAuthActivityIntent(): Intent
+}
