+++ core/network/src/main/java/org/sopt/official/network/service/RefreshService.kt
@@ -5,7 +5,7 @@ import org.sopt.official.network.model.request.RefreshRequest
 import retrofit2.http.Body
 import retrofit2.http.PATCH
 
-interface AuthService {
+interface RefreshService {
     @PATCH("auth/refresh")
     suspend fun refresh(
         @Body body: RefreshRequest
